add Dockerfile
This uses Python version 3.11 with Debian slim-bullseye (11).

I tried for a quite while to build it on Alpine, but Zenroom bindings
fail to work no matter what I do

It uses Uvicorn that listens on TCP over IPv4 using port 8000.

A process manager such as Gunicorn was excluded for simplicity and
can be used separetely, or orchestration can be used instead.